+++ backend/micro-gerenciamento/db.py
@@ -1,0 +1,13 @@
+import mysql.connector
+from dotenv import load_dotenv
+import os
+
+load_dotenv()
+
+def get_connection():
+    return mysql.connector.connect(
+        host=os.getenv("DB_HOST"),
+        user=os.getenv("DB_USER"),
+        password=os.getenv("DB_PASSWORD"),
+        database=os.getenv("DB_NAME")
+    )
